test: split tile tests into individual files
