+++ examples/spring-example/src/test/kotlin/com/abbott/mosaic/examples/spring/orders/tile/AddressTileTest.kt
@@ -1,0 +1,17 @@
+package com.abbott.mosaic.examples.spring.orders.tile
+
+import com.abbott.mosaic.examples.spring.orders.OrderRequest
+import com.abbott.mosaic.examples.spring.orders.model.Address
+import com.abbott.mosaic.test.TestMosaicBuilder
+import kotlinx.coroutines.runBlocking
+import org.junit.jupiter.api.Test
+
+class AddressTileTest {
+  @Test
+  fun `address tile uses request id`() =
+    runBlocking {
+      val testMosaic = TestMosaicBuilder().withRequest(OrderRequest("order-1")).build()
+      val expected = Address("123 Main St", "Springfield")
+      testMosaic.assertEquals(AddressTile::class, expected)
+    }
+}
