+++ examples/spring-example/src/test/kotlin/com/abbott/mosaic/examples/spring/orders/tile/CarrierQuotesTileTest.kt
@@ -1,0 +1,23 @@
+package com.abbott.mosaic.examples.spring.orders.tile
+
+import com.abbott.mosaic.examples.spring.orders.model.Address
+import com.abbott.mosaic.examples.spring.orders.model.Quote
+import com.abbott.mosaic.test.TestMosaicBuilder
+import kotlinx.coroutines.runBlocking
+import org.junit.jupiter.api.Test
+
+class CarrierQuotesTileTest {
+  @Test
+  fun `carrier quotes tile uses address tile`() =
+    runBlocking {
+      val address = Address("123 Main St", "Springfield")
+      val carriers = listOf("UPS", "FEDEX")
+      val quotes =
+        mapOf(
+          "UPS" to Quote("UPS", 5.99),
+          "FEDEX" to Quote("FEDEX", 7.49),
+        )
+      val testMosaic = TestMosaicBuilder().withMockTile(AddressTile::class, address).build()
+      testMosaic.assertEquals(CarrierQuotesTile::class, carriers, quotes)
+    }
+}
